Big overhaul to the sized_number library - create a trait, move everything to use that trait, and write some tests
