+++ src/sized_number/binary_options/mod.rs
@@ -1,7 +1,10 @@
 use std::fmt::Binary;
 use std::mem;
 
+use simple_error::{SimpleResult, bail};
 use serde::{Serialize, Deserialize};
+
+use crate::sized_number::SizedOptions;
 
 /// Configure display options for [`SizedDisplay::Binary`]
 #[derive(Debug, Clone, Copy, Serialize, Deserialize)]
@@ -54,3 +57,101 @@ impl BinaryOptions {
     }
 }
 
+impl SizedOptions for BinaryOptions {
+    fn to_s_i8(&self, v:   i8)   -> SimpleResult<String> {
+        Ok(self.to_s(Box::new(v)))
+    }
+
+    fn to_s_i16(&self, v:  i16)  -> SimpleResult<String> {
+        Ok(self.to_s(Box::new(v)))
+    }
+
+    fn to_s_i32(&self, v:  i32)  -> SimpleResult<String> {
+        Ok(self.to_s(Box::new(v)))
+    }
+
+    fn to_s_i64(&self, v:  i64)  -> SimpleResult<String> {
+        Ok(self.to_s(Box::new(v)))
+    }
+
+    fn to_s_i128(&self, v: i128) -> SimpleResult<String> {
+        Ok(self.to_s(Box::new(v)))
+    }
+
+
+    fn to_s_u8(&self, v:   u8)   -> SimpleResult<String> {
+        Ok(self.to_s(Box::new(v)))
+    }
+
+    fn to_s_u16(&self, v:  u16)  -> SimpleResult<String> {
+        Ok(self.to_s(Box::new(v)))
+    }
+
+    fn to_s_u32(&self, v:  u32)  -> SimpleResult<String> {
+        Ok(self.to_s(Box::new(v)))
+    }
+
+    fn to_s_u64(&self, v:  u64)  -> SimpleResult<String> {
+        Ok(self.to_s(Box::new(v)))
+    }
+
+    fn to_s_u128(&self, v: u128) -> SimpleResult<String> {
+        Ok(self.to_s(Box::new(v)))
+    }
+
+    fn to_s_f32(&self, _v:  f32) -> SimpleResult<String> {
+        bail!("Floating point values cannot be displayed as binary")
+    }
+
+    fn to_s_f64(&self, _v:  f64) -> SimpleResult<String> {
+        bail!("Floating point values cannot be displayed as binary")
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    use pretty_assertions::assert_eq;
+    use simple_error::SimpleResult;
+    use crate::sized_number::{Context, SizedDisplay, SizedDefinition};
+
+    #[test]
+    fn test_binary_i8() -> SimpleResult<()> {
+        let data = b"\x00\x00\x12\xab\xFF\xFF\xFF\xFF".to_vec();
+
+        let tests = vec![
+            // index    prefix      padded   expected
+            (   0,      true,       true,    "0b00000000"),
+            (   1,      true,       true,    "0b00000000"),
+            (   2,      true,       true,    "0b00010010"),
+            (   3,      true,       true,    "0b10101011"),
+            (   4,      true,       true,    "0b11111111"),
+            (   5,      true,       true,    "0b11111111"),
+
+            (   0,      false,      false,   "0"),
+            (   1,      false,      false,   "0"),
+            (   2,      false,      false,   "10010"),
+            (   3,      false,      false,   "10101011"),
+            (   4,      false,      false,   "11111111"),
+            (   5,      false,      false,   "11111111"),
+        ];
+
+        for (index, padded, prefix, expected) in tests {
+            let context = Context::new_at(&data, index);
+
+            assert_eq!(
+                expected,
+                SizedDefinition::U8.to_string(
+                    context,
+                    SizedDisplay::Binary(BinaryOptions {
+                        prefix: prefix,
+                        padded: padded,
+                    })
+                )?
+            );
+        }
+
+        Ok(())
+    }
+}
